requirements.txt
streamlit>=1.30.0
streamlit-folium>=0.15.0
folium>=0.15.0
streamlit-extras>=0.3.0
sentinelhub>=3.10.0
pystac-client>=0.7.0
stackstac>=0.5.0
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.0
numpy>=1.24.0
pandas>=2.1.0
scikit-image>=0.22.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
plotly>=5.17.0
arxiv>=2.0.0
aiohttp>=3.9.0
asyncio-contextmanager>=1.0.0
